agrego confirmacion al borrar una materia y cambio la posicion de renderizado de los errores

+++ src/components/configuration/Course.js
@@ -39,35 +39,29 @@ export default function Course(){
     }
 
     const _deleteCourse = ()=>{
-     
-        deleteCourse(codigo).then((response)=>{
-            if(response.status == 204 || response.status == 200 || response.status == 201){
-                setMessage('Materia eliminada exitosamente');
-                setShowMessage(true);
-                setCallError(false);
-            }else{
-                setMessage(response.data.error+ ": " + response.data.message);
+        if (window.confirm("Estas seguro que deseas borrar la materia")) {
+            deleteCourse(codigo).then((response)=>{
+                if(response.status == 204 || response.status == 200 || response.status == 201){
+                    setMessage('Materia eliminada exitosamente');
+                    setShowMessage(true);
+                    setCallError(false);
+                }else{
+                    setMessage(response.data.error+ ": " + response.data.message);
+                    setShowMessage(true);
+                    setCallError(true);
+                }
+            }).catch((error)=>{
+                
+                setMessage(error.code+" : "+((error.response.data[0] !== undefined)? error.response.data[0].message : error.response.data.message));
                 setShowMessage(true);
                 setCallError(true);
-            }
-        }).catch((error)=>{
-            
-            setMessage(error.code+" : "+((error.response.data[0] !== undefined)? error.response.data[0].message : error.response.data.message));
-            setShowMessage(true);
-            setCallError(true);
-        })
-        
+            })
+        }
     }
 
     return(<>   
         <br></br>
         <br></br>
-        {
-          showMessage?
-            <><AlertRequest message={message} click={()=>{setShowMessage(false)}} show={showMessage} error={callError}></AlertRequest></>
-            :
-            <></>
-        }
 
         <Container>
         <Accordion>
@@ -75,7 +69,12 @@ export default function Course(){
                 <Accordion.Header>Crear Materia</Accordion.Header>
                 <Accordion.Body>
                 <Form>
-
+                    {
+                    showMessage?
+                        <><AlertRequest message={message} click={()=>{setShowMessage(false)}} show={showMessage} error={callError}></AlertRequest></>
+                        :
+                        <></>
+                    }
                     <Form.Group className="mb-3" controlId="nombControl">
                         <Form.Label>Nombre: </Form.Label>
                         <Form.Control type="text" placeholder="Nombre de la Materia" onChange={(e)=>{setNombre(e.target.value)}} />
@@ -108,12 +107,17 @@ export default function Course(){
             <Accordion.Item eventKey="1">
                 <Accordion.Header>Borrar Materia</Accordion.Header>
                 <Accordion.Body>
+                {
+                    showMessage?
+                        <><AlertRequest message={message} click={()=>{setShowMessage(false)}} show={showMessage} error={callError}></AlertRequest></>
+                        :
+                        <></>
+                }
                 <Form.Group className="mb-3" controlId="codControl">
-                        <Form.Label>Codigo de Materia: </Form.Label>
-                        <Form.Control type="text" placeholder="Codigo de la meteria a eliminar" onChange={(e)=>{setCodigo(e.target.value)}} />
-                    </Form.Group>
-                      <Button variant="outline-danger" onClick={_deleteCourse}>Borrar</Button>
-
+                    <Form.Label>Codigo de Materia: </Form.Label>
+                    <Form.Control type="text" placeholder="Codigo de la meteria a eliminar" onChange={(e)=>{setCodigo(e.target.value)}} />
+                </Form.Group>
+                    <Button variant="outline-danger" onClick={_deleteCourse}>Borrar</Button>
                 </Accordion.Body>
             </Accordion.Item>
             <Accordion.Item eventKey="2">
